refactor(metadata): 👷 personal - krsiak and crypto
